Catches SyntaxError when parsing JSON
Adds a try/catch block when parsing a mod's JSON file. If a syntax error in the JSON is encountered, this will print to the console that it has been skipped instead of potentially halting further JavaScript.
Additionally adds a check for all required JSON fields. This skips non-compliant files.